refactor(shared): 归还领域解析器所有权
