+++ src/cli/lib/run-core-evaluation.ts
@@ -27,7 +27,7 @@ import {
 } from '../../eval-workflows/downstream-projections/index.js';
 import { discoverBatchSkills } from '../../inputs/skill-loader.js';
 import { projectReportsDir, globalReportsDir } from '../../measurement-artifacts/directories.js';
-import { generateRunId } from '../../shared/run-id.js';
+import { generateRunId } from '../../measurement-artifacts/run-id.js';
 import type { RunConfig } from './parse-run-config.js';
 import type { CliLang } from './i18n.js';
