+++ src/doctor/rules.ts
@@ -26,7 +26,7 @@ import yaml from 'js-yaml';
 import { tDoctorMessage } from './messages.js';
 import type { Lang } from '../shared/language.js';
 import { preflightDependencies } from '../preflight/dependencies.js';
-import { validateSkillHardRules, validateSkillWorkflows } from '../shared/hard-rules.js';
+import { validateSkillHardRules, validateSkillWorkflows } from '../skill-definition/hard-rules.js';
 import type { DependencyIssue } from '../preflight/dependencies.js';
 import type {
   DoctorRule,
