+++ src/observability/experience-frontmatter.ts
@@ -22,7 +22,7 @@
  */
 
 import { readFileSync } from 'node:fs';
-import { parseSkillFrontmatter, validateSkillHardRules, validateSkillWorkflows } from '../shared/hard-rules.js';
+import { parseSkillFrontmatter, validateSkillHardRules, validateSkillWorkflows } from '../skill-definition/hard-rules.js';
 import { findSkillMdPath } from './skill-chain.js';
 import type { ExperienceRuntimeSkillType } from './contracts/experience.js';
