+++ src/studio/http/routes/knowledge.ts
@@ -25,7 +25,7 @@ import {
   type SkillHealthReport,
 } from '../../../observability/skill-health-analyzer.js';
 import { parseSkillHealthReport } from '../../../observability/skill-health-report.js';
-import { parseDoctorReport } from '../../../shared/doctor-report.js';
+import { parseDoctorReport } from '../../../doctor/report-parser.js';
 import type { Lang } from '../../../shared/language.js';
 import { buildSkillIndex } from '../../application/index.js';
 import { renderDoctorDetail } from '../../presentation/doctor-detail-renderer.js';
